requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.1
websockets==12.0
pydantic==2.5.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
lxml==4.9.3
aiohttp==3.9.1
newspaper3k==0.2.8
google-generativeai==0.3.1
finnhub-python==2.4.19
python-dateutil==2.8.2
yfinance==0.2.33
numpy==1.24.3
pandas==2.0.3